requirements.txt
#installing discord.py
discord.py

#for GEMINI API
google

google-genai